Added BOOST_WINDOWS_API support to parse.cpp
path.c_str() returns value_type*, which is a wchar_t* if BOOST_WINDOWS_API is defined. Tinyxml uses fopen to open files, which does not take wchar_t argument. This fix leaves tinyxml invariant, but there can be an error if a file with non-ascii characters is being opened. Windows workaround is wfopen, but this function is not available on Cygwin, which is the environment that is most likely to be used on Windows to build sdpb. Cygwin follows posix, so that's why wfopen is not available. However, boost does not believe that cygwin is posix and defines BOOST_WINDOWS_API.