Serialization tests and benchmarks for read/write SDP block_data_XXX.bin and .json
You can run benchmark as
./build/unit_tests [!benchmark] --benchmark-samples 10 --benchmark-no-analysis

On my laptop, binary format is ~20x-30x faster than JSON for combined write+read to an in-memory stringstream.
If disk IO is a bottleneck (which sounds more realistic), one can expect more moderate speedup, ~2.4x because binary format is more compact than JSON.

This is a part of #111 Benchmarks for reading/writing sdp.zip.