~Shared_Window_Array(): Disable MPI synchronization during stack unwinding.
If one rank throws an exception and another one doesn't, the first rank will hang forever on the window fence.
Thus, we disable this fence and assume that the program will abort.
NB: if exception is caught after that and program continues working, it will probably hang on the next synchronization point!