Fix #175 Debug mode: write profiling for actual run
After actual run, we write profiling data to ck.profiling/ and block_timings (for the last step) to ck/block_timings
TODO: do not write block_timings if number of iterations == 1?

Old data is moved to ck.profiling.0/ (or ck.profiling.1/ etc., if .0 exists) and block_timings.0

Previously, we wrote data only for the timing run.