Test_Case_Runner: write both stdout and stderr to a single log file
Also we write stdout synchronously, which saves us from saving all output to the large buffer (e.g. if --verbosity=debug)

TODO: now stderr is appended to the end of log. Can we set up redirection to (1) write stdout+stderr to the same file, (2) at the same time, write stderr to a string?