Increase automatic --maxSharedMemory limit from 40% to 50% of remaining memory.
When we have a lot of RAM, 50% is enough.
In extreme cases (stress-tensors-3d nmax=18, 4 or 5 nodes), 40% doesn't prevent OOM, we need to reduce e.g. to 25% or even lower.
This is too small for most of the cases, we don't want to be too restrictive.

Since we cannot guarantee it anyway, we'll simply choose a nice number, 50%.
In case of OOM user should adjust the limit manually.